pulls

+++ server/source/platinum-cms/src/platinum/cms/admin/web/rest/PostAdminResource.java
@@ -18,8 +18,11 @@ import org.codehaus.jettison.json.JSONException;
 import org.codehaus.jettison.json.JSONObject;
 
 import platinum.cms.admin.service.PostAdminManager;
+import platinum.cms.runtime.service.PostRuntimeManager;
 import platinum.cms.common.PostStatus;
+import platinum.cms.common.PostType;
 import platinum.cms.common.entity.PostEntity;
+import platinum.cms.common.entity.CallBackEntity;
 import platinum.cms.common.search.PostSearchEngine;
 import platinum.cms.common.search.PostSearchResult;
 import platinum.cms.common.serialization.PostJSONSerializer;
@@ -35,12 +38,33 @@ public class PostAdminResource extends AbstractResource
 	public Response loadPosts(
 			@QueryParam("categoryId") String p_categoryId,
 			@QueryParam("keywords") String p_keywords,
+			@QueryParam("dashboardhitcount") String dashboardhitcount,
+			@QueryParam("dashboardtime") String dashboardtime,
 			@QueryParam("pageIndex") @DefaultValue("0") int p_pageIndex,
-			@QueryParam("pageSize") @DefaultValue("20") int p_pageSize
+			@QueryParam("pageSize") @DefaultValue("10") int p_pageSize
 			) throws JSONException
-	{
+	{   
 		List<PostEntity> posts = null;
+		 String _CurrentUser=	  Membership.getInstance().getCurrentUser().getUserRole();
 		
+		 
+		 
+		 if(dashboardtime!=null)
+		 {  
+			
+					posts=PostRuntimeManager.getInstance().loadLatestPostsByCategory(dashboardtime, false, null, p_pageIndex, p_pageSize);
+					return responseWithJSONArray(PostJSONSerializer.toSimpleArray(posts));
+			
+			  
+		 }
+		
+		 if(dashboardhitcount!=null)
+		 {  
+			 posts=PostRuntimeManager.getInstance().loadTopPostsByCategory("news", p_pageSize);
+			return responseWithJSONArray(PostJSONSerializer.toSimpleArray(posts));
+		 }
+		
+		 
 		if (p_keywords != null)
 		{
 			if (p_keywords.length() == 32)
@@ -54,16 +78,31 @@ public class PostAdminResource extends AbstractResource
 			}
 			if (posts == null)
 			{
-				List<PostSearchResult> docs = PostSearchEngine.getInstance().search(p_keywords);
+				List<PostSearchResult> docs = PostSearchEngine.getInstance().search(p_keywords, p_pageIndex, 1000, 500, 200);
 				return responseWithJSONArray(PostJSONSerializer.toSimpleArray2(docs));
 			}
 		}
-		else
+		else  if(_CurrentUser.equals("Administrator"))
+		{    
+		
+		
+			   posts=PostAdminManager.getInstance().loadPostsByCategory(null, p_pageIndex, p_pageSize);
+	
+		  
+		   
+			
+			
+		}
+		else 
 		{
-			posts = PostAdminManager.getInstance().loadPostsByCategory(null, p_pageIndex, p_pageSize);
+			posts=PostRuntimeManager.getInstance().loadLatestPostsByCategory(_CurrentUser, false, null, p_pageIndex, p_pageSize);
+			
 		}
 		return responseWithJSONArray(PostJSONSerializer.toSimpleArray(posts));
 	}
+	
+
+	
 	
 	@GET
 	@Path("/{id}")
@@ -72,6 +111,19 @@ public class PostAdminResource extends AbstractResource
 			) throws JSONException
 	{
 		PostEntity post = PostAdminManager.getInstance().getPostById(p_id);
+		
+		if((post.getCategoryId()).equals(Membership.getInstance().getCurrentUser().getUserRole()))
+		{
+			 post = PostAdminManager.getInstance().getPostById(p_id);
+		}
+		else
+		{
+			post=null;
+		}
+		if("Administrator".equals( Membership.getInstance().getCurrentUser().getUserRole() ))
+		{
+			 post = PostAdminManager.getInstance().getPostById(p_id);
+		}	
 		
 		String ifModifiedSince = getHttpHeader("If-Modified-Since");
 		if (ifModifiedSince != null)
@@ -97,7 +149,16 @@ public class PostAdminResource extends AbstractResource
 		
 		_parsePostFromJSON(postJSON, post);
 		
-		PostAdminManager.getInstance().savePost(post);
+		String _postCategoryId=post.getCategoryId();
+	
+		if(_postCategoryId.equals(Membership.getInstance().getCurrentUser().getUserRole()))
+		{
+			PostAdminManager.getInstance().savePost(post);
+		}
+		if("Administrator".equals( Membership.getInstance().getCurrentUser().getUserRole() ))
+		{
+			PostAdminManager.getInstance().savePost(post);
+		}
 		
 		JSONObject jsonResult = _generateSimplePostJSON(post);
 		return responseWithJSONObject(jsonResult);
@@ -117,8 +178,9 @@ public class PostAdminResource extends AbstractResource
 			
 			_parsePostFromJSON(postJSON, post);
 			
-			PostAdminManager.getInstance().updatePost(post);
-			
+		
+				PostAdminManager.getInstance().updatePost(post);
+				
 			JSONObject jsonResult = _generateSimplePostJSON(post);
 			return responseWithJSONObject(jsonResult);
 		}
@@ -127,6 +189,27 @@ public class PostAdminResource extends AbstractResource
 			return responseWithException("没有找到标识为“" + p_id + "”的文章。");
 		}
 	}
+	@PUT
+	@Path("nihao")
+	public Response callbacksave(
+			@FormParam("post") String p_postJSONString
+			) throws JSONException
+		{   
+			
+			JSONObject postJSON = new JSONObject(p_postJSONString);
+			
+			CallBackEntity post=new CallBackEntity();
+			post.setName(postJSON.getString("name"));
+			post.setTitle(postJSON.getString("title"));
+			post.setContent(postJSON.getString("content"));
+			
+			PostAdminManager.getInstance().savePost(post);
+			return null;
+			
+			
+			
+			
+		}
 
 	
 	@DELETE
@@ -134,14 +217,22 @@ public class PostAdminResource extends AbstractResource
 	public Response deletePost(
 			@PathParam("id") String p_id
 			) throws JSONException
-	{
-		PostAdminManager.getInstance().deletePost(p_id);
+	{    
+		PostEntity post = PostAdminManager.getInstance().getPostById(p_id);
+		
+		
+		if((post.getCategoryId()).equals(Membership.getInstance().getCurrentUser().getUserRole()))
+		
+		{
+			PostAdminManager.getInstance().deletePost(p_id);
+		}
+		
+		if("Administrator".equals( Membership.getInstance().getCurrentUser().getUserRole() ))
+		{
+			PostAdminManager.getInstance().deletePost(p_id);
+		}	
 		return responseOK();
 	}
-	
-	
-	
-	
 	
 	
 	
@@ -162,6 +253,14 @@ public class PostAdminResource extends AbstractResource
 		{
 			post.setSubcategoryId(null);
 		}
+		if (postJSON.getString("homeSubcategoryId") != "null")
+		{
+			post.setHomeSubcategoryId(postJSON.getString("homeSubcategoryId"));
+		}
+		else
+		{
+			post.setHomeSubcategoryId(null);
+		}
 		post.setSource(postJSON.getString("source"));
 		if (postJSON.getString("photoURL") != "null")
 		{
@@ -172,6 +271,7 @@ public class PostAdminResource extends AbstractResource
 			post.setPhotoURL(null);
 		}
 		post.setPostStatus(PostStatus.values()[postJSON.getInt("postStatus")]);
+		post.setPostType(PostType.values()[postJSON.getInt("postType")]);
 	}
 	
 	private JSONObject _generateSimplePostJSON(PostEntity post)
